Does some refactor, fixes some changes
