+++ app/src/main/java/com/example/tmdbapp/domain/model/Movie.kt
@@ -1,10 +1,7 @@
 package com.example.tmdbapp.domain.model
 
-import androidx.annotation.DrawableRes
-
 data class Movie(
     val movieTitle: String,
-    @DrawableRes
     val moviePoster: Int,
     val movieRating: Int
 )
