+++ app/src/main/java/com/example/tmdbapp/presentation/home/HomeScreen.kt
@@ -7,6 +7,13 @@ import com.example.tmdbapp.core.utils.Constants
 @Composable
 fun HomeScreen(){
     Column {
+        SearchBar(
+            value = "",
+            placeholder = "Search for the movie",
+            isError = false,
+            onSearch = { /*TODO*/ },
+            onTextChanged = {}
+        )
         TopMovies(topMovies = Constants.getMovies())
         TopMovies(topMovies = Constants.getMovies())
     }
