+++ app/src/main/java/com/example/tmdbapp/presentation/home/MoviePoster.kt
@@ -1,17 +1,14 @@
-package com.example.tmdbapp.presentation.splash
+package com.example.tmdbapp.presentation.home
 
-import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.tmdbapp.R
 import com.example.tmdbapp.ui.theme.TMDBAppTheme
 
 @Composable
 fun MoviePoster(
     // Before API calls, later change to URL
-    @DrawableRes
     movie_Poster: Int
 ){
     Image(
@@ -23,6 +20,6 @@ fun MoviePoster(
 @Composable
 fun MoviePosterPreview(){
     TMDBAppTheme {
-        MoviePoster(movie_Poster = R.drawable.spiderman)
+        //MoviePoster(movie_Poster = R.drawable.spiderman)
     }
 }
